inputhook: move inputhook_qt4 related code in own file
